+++ Config/Config.cfg.php
@@ -1,0 +1,3 @@
+<?php abstract class Config {
+public abstract static function init();
+}#